+++ src/components/cards/ItemCard.component.jsx
@@ -61,7 +61,6 @@ function ItemCard({
                 item={item}
                 handleAddToShoppingList={handleAddToShoppingList}
                 handleUpdateItemQuantity={handleUpdateItemQuantity}
-                isRenderedInShoppingCart={handleUpdateItem}
               />
             )}
           </>